+++ AIO Clipboard & Search/App.xaml.cs
@@ -1,5 +1,7 @@
+using AIO_Hybrid_Clipboard.Services;
 using System;
 using System.Threading;
+using System.Threading.Tasks;
 using System.Windows;
 
 namespace AIO_Hybrid_Clipboard
@@ -15,6 +17,8 @@ namespace AIO_Hybrid_Clipboard
 
         protected override void OnStartup(StartupEventArgs e)
         {
+            RegisterGlobalExceptionLogging();
+
             _mutex = new Mutex(true, MutexName, out _ownsMutex);
             if (!_ownsMutex)
             {
@@ -49,6 +53,26 @@ namespace AIO_Hybrid_Clipboard
             })
             { IsBackground = true, Name = "SecondInstanceListener" };
             listener.Start();
+        }
+
+        // Last-resort logging: without these hooks an unexpected exception kills
+        // the tray app silently and leaves no trace in AIO_Cache/app.log.
+        private void RegisterGlobalExceptionLogging()
+        {
+            DispatcherUnhandledException += (_, args) =>
+            {
+                Log.Error("Unhandled UI exception", args.Exception);
+                args.Handled = true; // keep the tray app alive; state-corrupting bugs still surface in the log
+            };
+
+            AppDomain.CurrentDomain.UnhandledException += (_, args) =>
+                Log.Error("Unhandled domain exception (terminating)", args.ExceptionObject as Exception);
+
+            TaskScheduler.UnobservedTaskException += (_, args) =>
+            {
+                Log.Error("Unobserved task exception", args.Exception);
+                args.SetObserved();
+            };
         }
 
         protected override void OnExit(ExitEventArgs e)